+++ pkg/integration/tests/test_list.go
@@ -137,12 +137,12 @@ var tests = []*components.IntegrationTest{
 	interactive_rebase.SwapInRebaseWithConflictAndEdit,
 	interactive_rebase.SwapWithConflict,
 	misc.ConfirmOnQuit,
+	misc.CopyToClipboard,
 	misc.InitialOpen,
 	misc.RecentReposOnLaunch,
 	patch_building.Apply,
 	patch_building.ApplyInReverse,
 	patch_building.ApplyInReverseWithConflict,
-	patch_building.CopyPatchToClipboard,
 	patch_building.MoveToEarlierCommit,
 	patch_building.MoveToEarlierCommitNoKeepEmpty,
 	patch_building.MoveToIndex,
